Refactored our Coordinates and outdated design
Yes, I already have outdated designs. The PathTranslator and
Translatable are outdated. It felt bad and convoluted for such a small
use case. The goal was to simplify how pieces move on a board with
equations but this is stupid to do in an OOP language that already has
lambda functions. This also jeopardizes the Model-View-Controller
paradigm as, thus far, it's set that each "translatable" would need to
be attached to an icon or model. I could have the controller
micro-manage these, but these objects act as controllers in and of
themselves, operating without care of what kind of model or view is
being used as well as not caring what other translatables may exist.
However, that's impossible with some games like chess, where the
position of other pieces influence where another piece can move, make
such a selfish design impractical and impossible. This works with
Monopoly, but not with other games and it's for other games I'd like to
open my design to. As for Coordinates, simple, I found a Java class that
already did what I wanted. I'm not too happy with how I have to call
them each time but I'll look into that.