Refactoring with optimizations and clarity code
First off, I changed the name of RelativeDimensions to Relativedim so I
wouldn't have to type so much. In addition, I made the RelativePoint
class which is just a wrapper to Point2D.Double because, one again, it's
less annoying to type out. I also refactored StickerPages to render in,
at best, O(1) time instead of always the sum of the area occupied by all
stickers with some prerendered images of the stickers. And lastly, I
refactored BufferedRender to make it's instance variables private even
to it's children and instead I'm using getters and setters since, prior
to this refactor, any other BufferedRender child class could access
these members in other BufferendRender objects. e.g. StickerPages could
directly change/interact with the dirty value of it's Stickers. This was
not ok especially for any future implementations which may want to
override these methods