Huffman Tree save files started
Currently having trouble working with binary. Either encoding the path
of a given character into the binary output stream is failing or writing
the binary output stream to the file is failing. The earlier would make
more sense but I'm concerned about endiness for the output affecting the
resulting binary